refactor: extract shared routes projector into @webjsdev/mcp
Pull the list_routes projection into a leaf routes-report.js module,
mirroring check-report.js. The MCP list_routes tool now delegates to
projectRoutes(), and the same projector is exported as
`@webjsdev/mcp/routes-report` so the upcoming `webjs routes --json`
command can reuse it and stay byte-identical to the MCP tool. No
behaviour change: list_routes output is unchanged.
